Port PlatformerGameMode to Haxe
The GameMode class defines the game being played, and enforces the game rules.
By starting from a central class like this, we needed to declare some `Externs` so we
could reference other C++ code that still isn't ported to Haxe. These externs are found in
the `Haxe/Externs` directory.

Also note that in Haxe we don't have to use the same naming structure as C++. In fact, using
a package is many times required, while the exported C++ name can be defined with the @:uname metadata.

There are a fair amount of `@:uexpose` functions and variables on this code. They are there
so the other C++ code can call these functions. As we advance on the port, we will be able
to take off these `@:uexpose` functions - and use more Haxe-like features, like Arrays, getters and setters
(instead of `TArrays`, `Get*`, `Set*` functions)

Lasty, note that unlike C++, struct variables that aren't declared as `@:uexpose` need to be initialized
(see GameMode.CurrentTimes vs GameMode.BestTimes)