Add ant build script to simplify building and testing from the command-line
In order for it to work, you need to define the UE4 environment variable.
You can always use the editor compilation instead, or issue the commands yourself.
This script however documents how you would do it through the command-line, which is sometimes
very useful for debugging some compilation issues - specially if this is your first compilation