Port PlatformerPlayerController
This file was very straight-forward to port. There is one noteworthy case here:
Even though we ported the PlayerController code to Haxe, we still kept an `External`
reference in the `Externs` directory.
We did that so that we could correctly type the `FPlatformerIngameMenu` extern.
Since the extern baking happens in an earlier compilation phase, the extern baker isn't able
to have access to the code inside the `Scripts` or `Static` directories, so if you want to reference
anything that is there, you'll need to still create an extern for them.